pyproject.toml
[tool.poetry]
name = "opa-permission-api"
version = "1.0.0"
description = "FastAPI backend server that integrates with OPA for permission management"
authors = ["Sylvain <zslvain@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.109.0"
uvicorn = {extras = ["standard"], version = "0.27.0"}
sqlalchemy = "2.0.25"
alembic = "1.13.1"
boto3 = "1.34.34"
httpx = "0.26.0"
pydantic-settings = "2.1.0"
psycopg2-binary = "2.9.9"
python-multipart = "0.0.6"
python-jose = {extras = ["cryptography"], version = "3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
